Refactor application structure and add language toggle component with styles
