+++ LTU.SearchEngine.Frontend/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './css/index.css'
 import App from './App.tsx'
 import { BrowserRouter } from 'react-router-dom'
